pyproject.toml
[project]
name = "Freela_Backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "kauabarros-24", email = "martinsbarroskaua@gmail.com"},
    {name = "", email = ""},
    {name = "kaubarros-24", email = "martinsbarroskaua@gmail.com"},
]
dependencies = [
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.6.0",
    "djangorestframework-simplejwt>=5.3.1",
    "simplejwt>=2.0.1",
    "Pillow>=10.4.0",
    "WeasyPrint>=62.3",
    "reportlab>=4.2.2",
    "cloudinary>=1.41.0",
    "requests>=2.32.3",
    "django-filter>=24.3",
    "pdm>=2.20.1",
    "django-cloudinary-storage>=0.3.0",
    "whitenoise>=6.8.2",
    "psycopg2-binary>=2.9.10",
    "dj-database-url>=2.3.0",

]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.scripts]
r = "python3 manage.py runserver"
makemigrations = "python3 manage.py makemigrations"
migrate = "python3 manage.py migrate"
newMasterMan ="python3 manage.py createsuperuser"
newApp = "django-admin startapp"
newProject = "django-admin startproject"
shellp = "python3 manage.py shell"